updated python script for automatically managing the data from different boards and writing results to files - added resulting files from test runs

+++ ESP32Dev/Wine/src/main.cpp
@@ -207,8 +207,8 @@ void loop()
       String board = "\"board\":\"esp32dev\",";
       String model = "\"model\":\"wine\",";
       String result = "\"result\":" + String(resultClass) + ",";
-      String iteration = "\"iteration\":" + String(uint16_t(currentIteration)) + ",";
-      String time = "\"microseconds\":" + String(uint16_t(end));
+      String iteration = "\"iteration\":" + String(int(currentIteration)) + ",";
+      String time = "\"microseconds\":" + String(int(end));
       String endPar = "}";
 
       String resultString = startPar + board + model + result + iteration + time + endPar;
