+++ ESP32Wemos/Wine/src/main.cpp
@@ -200,15 +200,14 @@ void loop()
       Serial.println(" actual");
       Serial.println("Evaluation time: " + String(end) + " microseconds");
 
-      // DTO to be created
       //{"board": "esp32wemos", "model": "wine", "result": 1, "iteration": 1, "microseconds": 120}
 
       String startPar = "{";
       String board = "\"board\":\"esp32wemos\",";
       String model = "\"model\":\"wine\",";
       String result = "\"result\":" + String(resultClass) + ",";
-      String iteration = "\"iteration\":" + String(uint16_t(currentIteration)) + ",";
-      String time = "\"microseconds\":" + String(uint16_t(end));
+      String iteration = "\"iteration\":" + String(int(currentIteration)) + ",";
+      String time = "\"microseconds\":" + String(int(end));
       String endPar = "}";
 
       String resultString = startPar + board + model + result + iteration + time + endPar;
