+++ ESP8266/Sine/src/main.cpp
@@ -148,15 +148,14 @@ void loop()
       Serial.println(predicted);
       Serial.println("Evaluation time: " + String(end) + " microseconds");
 
-      // DTO to be created
       //{"board": "esp8266", "model": "sin", "result": 3.14, "iteration": 1, "microseconds": 120}
 
       String startPar = "{";
       String board = "\"board\":\"esp8266\",";
       String model = "\"model\":\"sin\",";
       String result = "\"result\":" + String(x) + ",";
-      String iteration = "\"iteration\":" + String(uint16_t(currentIteration)) + ",";
-      String time = "\"microseconds\":" + String(uint16_t(end));
+      String iteration = "\"iteration\":" + String(int(currentIteration)) + ",";
+      String time = "\"microseconds\":" + String(int(end));
       String endPar = "}";
 
       String resultString = startPar + board + model + result + iteration + time + endPar;
